méthodes pour la consultation des exécutions
git-svn-id: svn+ssh://svn.projects.openwide.fr/svnroot/owsi-tools/trunk/owsi-core@1167 d3474844-eb8e-4abc-b058-2b321fed648b

+++ owsi-core-components/owsi-core-component-jpa-more/src/main/java/fr/openwide/core/jpa/more/business/execution/dao/AbstractExecutionDaoImpl.java
@@ -15,7 +15,7 @@ import fr.openwide.core.jpa.more.business.execution.model.IExecutionType;
 import fr.openwide.core.jpa.more.business.execution.model.QAbstractExecution;
 
 public abstract class AbstractExecutionDaoImpl<E extends AbstractExecution<E, ?>>
-		extends GenericEntityDaoImpl<Integer, E> implements IAbstractExecutionDao<E> {
+		extends GenericEntityDaoImpl<Long, E> implements IAbstractExecutionDao<E> {
 
 	@Override
 	public List<E> list() {
